requirements.txt
google-cloud-speech
autopep8
pylint
pytest
